typedColumns know their length now
Their interface has been extended to know about the length of each
column chunk, this will be useful when preallocating bitmaps and other
auxiliary data structures.

+++ src/typed_column_test.go
@@ -71,7 +71,7 @@ func TestBasicStringColumn(t *testing.T) {
 		// TODO: this is the only test with roundtrips, because we don't have nth value implemented anywhere else
 		// that's because we would have to have interface{} as the return value, and that's no good for individual values
 		for j, val := range vals {
-			got := nc.nthValue(int64(j))
+			got := nc.nthValue(uint32(j))
 			if got != val {
 				t.Errorf("expecting %v, got %v", val, got)
 				return
